Implementação de métodos de segurança para validação de transações

+++ src/main/java/com/ucan/skawallet/back/end/skawallet/repository/TransactionRepository.java
@@ -1,6 +1,6 @@
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
+ * Click nbfs://nbhost/SystemFileSystem/Templates/SpringFramework/Repository.java to edit this template
  */
 package com.ucan.skawallet.back.end.skawallet.repository;
 
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Repository;
  * @author azm
  */
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Integer>
+public interface TransactionRepository extends JpaRepository<Transaction, Long>
 {
+
 }
